+++ js/components/App/index.js
@@ -1,0 +1,8 @@
+/**
+ * @format
+ * @flow strict
+ */
+
+import { default as App } from './App';
+
+export default App;
