+++ js/constants/EstablishData.js
@@ -3,11 +3,9 @@
  * @flow strict
  */
 
-import type { EstablishData } from '../flow-typed/establish-data';
+import { ACCESS_ID, ENV, ENV_HOST, MERCHANT_ID, MERCHANT_REFERENCE } from '../../env';
 
-import { ACCESS_ID, ENV, ENV_HOST, MERCHANT_ID, MERCHANT_REFERENCE } from '../env';
-
-const establishData: EstablishData = {
+const EstablishData: Establish = {
   accessId: ACCESS_ID,
   cancelUrl: '#',
   currency: 'USD',
@@ -37,4 +35,4 @@ const establishData: EstablishData = {
   returnUrl: '#',
 };
 
-export default establishData;
+export default EstablishData;
